move compute_fiat_shamir_challenge into cryptography
